requirements.txt
# Image processing
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# Numerical computing
numpy>=1.24.0

# Progress bars
tqdm>=4.65.0

# Additional dependencies for refactored version
# Configuration management
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# CLI interface
click>=8.1.0

# Testing (dev dependency)
pytest>=7.4.0

# Type checking (dev dependency)  
mypy>=1.5.0

# Web interface
gradio>=4.0.0